+++ lib/capsum.rb
@@ -1,4 +1,5 @@
 require "capsum/version"
+require "capsum/patch_sshkit_force_encoding_to_utf8"
 
 module Capsum
 end